refactored routes to submodule
