+++ plastered/api/routes/__init__.py
@@ -1,0 +1,4 @@
+from plastered.api.routes.api_routes import plastered_api_router
+from plastered.api.routes.webserver_routes import plastered_web_router
+
+__all__ = ["plastered_api_router", "plastered_web_router"]
